Вынесена генерация sql в базовый класс; Улучшение обработки ошибок; типизация методов

+++ server/src/providers/registration/index.ts
@@ -1,7 +1,15 @@
-import {User, UserPassword} from '../common/entity/types'
-import RegistrationRepository from '../repositories/registration'
-import Hash from '../services/hash'
-import {Smtp} from '../services/smtp'
+import {User, UserPassword} from '../../common/entity/types'
+import RegistrationRepository from '../../repositories/registration'
+import Hash from '../../services/hash'
+import {Smtp} from '../../services/smtp'
+import {ApiError} from '../../errors'
+
+// Ошибка регистрации
+export class RegistrationError extends ApiError {
+    constructor(message: string, name: string = 'REGISTRATION_ERROR') {
+        super(message, name)
+    }
+}
 
 class RegistrationProvider {
     constructor(
@@ -12,20 +20,22 @@ class RegistrationProvider {
     }
 
     // Регистрация
-    signUp = async (user: User) => {
+    signUp = async (user: User): Promise<string> => {
         user.nickname = user.username
         user.username = user.username.toUpperCase()
         user.password = user.password.toUpperCase()
         user.password = this.hash.getHash(user.username, user.password)
         user.token = this.hash.getToken()
-        try {
-            await this.registrationRepository.selectAccountByQuery({
-                username: user.username,
-                email: user.email,
-            })
-            return Promise.reject('Ошибка регистрации, возможно логин или email уже занят')
-        } catch (e) {
+
+        // Если username и email свободен, то продолжить
+        const account = await this.registrationRepository.selectAccountByQuery({
+            username: user.username,
+            email: user.email,
+        })
+        if (account) {
+            throw new RegistrationError('Ошибка, возможно логин или email уже занят')
         }
+
         if (!parseInt(process.env.SEND_EMAIL as string)) {
             await this.registrationRepository.insertUserReg(user)
             await this.acceptReg(user.token)
@@ -34,51 +44,53 @@ class RegistrationProvider {
         return this.smtp.sendConfirmation(user.email, user.nickname, user.token).then(async () => {
             await this.registrationRepository.insertUserReg(user)
             return 'Письмо отправлено на почту'
-        }, () => {
-            return Promise.reject('Ошибка при отправке сообщения. Если ошибка повторится, то свяжитесь с администрацией')
         })
     }
 
     // Подтверждение регистрации
-    acceptReg = async (token: string) => {
+    acceptReg = async (token: string): Promise<void> => {
         const user = await this.registrationRepository.selectUserRegByToken(token)
-        try {
-            await this.registrationRepository.selectAccountByQuery({
-                username: user.username,
-                email: user.email,
-            })
-            return Promise.reject()
-        } catch (e) {
+
+        // Если username и email свободен, то продолжить
+        const account = await this.registrationRepository.selectAccountByQuery({
+            username: user.username,
+            email: user.email,
+        })
+        if (account) {
+            throw new RegistrationError('username или email занят')
         }
+
         const id = await this.registrationRepository.insertAccount(user)
         await this.registrationRepository.insertUser(user, id)
         return this.registrationRepository.removeUserReg(token)
     }
 
     // Подтверждение почты
-    acceptEmail = async (token: string) => {
+    acceptEmail = async (token: string): Promise<void> => {
         const user = await this.registrationRepository.selectUserEmailByToken(token)
-        try {
-            await this.registrationRepository.selectAccountByQuery({
-                email: user.email,
-            })
-            return Promise.reject()
-        } catch (e) {
+
+        // Проверка на занятость email
+        const account = await this.registrationRepository.selectAccountByQuery({
+            email: user.email,
+        })
+        if (account) {
+            throw new RegistrationError('email занят')
         }
         await this.registrationRepository.updateSecure(user)
         return this.registrationRepository.removeUserEmail(token)
     }
 
     // Редактирование настроек безопасноти
-    updateSecure = async (user: User) => {
+    updateSecure = async (user: User): Promise<string> => {
         user.token = this.hash.getToken()
-        try {
-            await this.registrationRepository.selectAccountByQuery({
-                email: user.email,
-            })
-            return Promise.reject('Ошибка, данный email возможно занят')
-        } catch (e) {
+
+        const account = await this.registrationRepository.selectAccountByQuery({
+            email: user.email,
+        })
+        if (account) {
+            throw new RegistrationError('Ошибка, данный email возможно занят')
         }
+
         if (!parseInt(process.env.SEND_EMAIL as string)) {
             await this.registrationRepository.insertAccountEmail(user)
             await this.acceptEmail(user.token)
@@ -87,13 +99,11 @@ class RegistrationProvider {
         return this.smtp.sendChangeEmail(user.email, user.token).then(async () => {
             await this.registrationRepository.insertAccountEmail(user)
             return 'Подтверждение отправлено на почту'
-        }, () => {
-            return Promise.reject('Ошибка при отправке сообщения. Если ошибка повториться, то свяжитесь с администрацией')
         })
     }
 
     // Редактирование пароля
-    updatePassword = (user: UserPassword) => {
+    updatePassword = (user: UserPassword): Promise<number> => {
         user.password = this.hash.getHash(user.username, user.password)
         return this.registrationRepository.updatePassword(user)
     }
